requirements.txt
# NaCCA School Management System - Dependencies

# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Security
Werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0

# PDF Generation
reportlab==4.0.7
weasyprint==60.1

# Utilities
gunicorn==21.2.0
python-dateutil==2.8.2
Pillow==10.1.0

# Form Validation
WTForms==3.1.1

# AI & Agentic Intelligence
groq==0.4.2
openai==1.6.1
requests==2.31.0
